ideal_response/tests/test_ssh_client.py
import unittest
import paramiko
from unittest.mock import patch, MagicMock
import unittest.mock
from ssh.ssh_client import SSHClient
class TestSSHClient(unittest.TestCase):
    @patch('paramiko.SSHClient')
    def setUp(self, mock_ssh):
        self.mock_ssh_instance = MagicMock()
        mock_ssh.return_value = self.mock_ssh_instance
        # self.ssh_client = SSHClient('test.example.com', 'testuser', 'password')

    def test_connect_success(self):
        # self.ssh_client.connect()
        # self.mock_ssh_instance.connect.assert_called_once_with(
        #     'test.example.com',
        #     username='testuser',
        #     password='password',
        #     key_filename=None
        # )
        pass

    def test_connect_auth_failure(self):
        # self.mock_ssh_instance.connect.side_effect = paramiko.AuthenticationException()
        # with self.assertRaises(paramiko.AuthenticationException):
        #     self.ssh_client.connect()
        pass

    def test_execute_command_success(self):
        # mock_stdin = MagicMock()
        # mock_stdout = MagicMock()
        # mock_stderr = MagicMock()
        # mock_stdout.channel.recv_exit_status.return_value = 0
        # mock_stdout.read.return_value = b'Command output'
        # self.mock_ssh_instance.exec_command.return_value = (mock_stdin, mock_stdout, mock_stderr)

        # result = self.ssh_client.execute_command('test command')
        # self.assertEqual(result, 'Command output')
        # self.mock_ssh_instance.exec_command.assert_called_once_with('test command')
        pass

    def test_execute_command_failure(self):
        # mock_stdin = MagicMock()
        # mock_stdout = MagicMock()
        # mock_stderr = MagicMock()
        # mock_stdout.channel.recv_exit_status.return_value = 1
        # mock_stderr.read.return_value = b'Command failed'
        # self.mock_ssh_instance.exec_command.return_value = (mock_stdin, mock_stdout, mock_stderr)

        # result = self.ssh_client.execute_command('test command')
        # self.assertIsNone(result)
        pass

if __name__ == '__main__':
    unittest.main()

# Contains the ssh test